requirements.txt
python-telegram-bot>=21.0
sqlalchemy>=1.4,<2.0
psycopg2-binary
